Supermarket sales data from Kaggle

Supermarket sales data from Kaggle

supermarket

Format

## `supermarket` A data frame with 1,000 rows and 17 columns:

Invoice.ID

Computer generated sales slip invoice identification number

Branch

Branch of supercenter (3 branches are available identified by A, B and C).

City

Location of supercenters

Customer.type

Type of customers, recorded by Members for customers using member card and Normal for without member card

Gender

Gender type of customer

Product.line

General item categorization groups

Unit.price

Price of each product in \$

Quantity

Number of products purchased by customer

Tax.5.

5% tax fee for customer buying

Total

Total price including tax

Date

Date of purchase

Time

Purchase time

Payment

Payment used by customer for purchase

cogs

Cost of goods sold

gross.margin.percentage

Gross margin percentage

gross.income

Gross income

Rating

Customer stratification rating on their overall shopping experience (On a scale of 1 to 10)

Source

<https://www.kaggle.com/datasets/aungpyaeap/supermarket-sales>